backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
telethon==1.34.0
sqlalchemy==2.0.29
apscheduler==3.10.4
python-dotenv==1.0.1
aiofiles==23.2.1
python-multipart==0.0.9
websockets==12.0
cryptg==0.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
